Rewrite the App logic
Unfortunately, my initial plan with the useFetch hook didn't work out. I had to rewrite the App component to use an useEffect fetch instead.

I also had to add a new helper function to handle the list operations. In order to handle the onDragEnd functionality.

There was a but with react-beautiful-dnd and UseStrict, I had to change libraries to an
updated one.